% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditModule.R
\name{addEditModuleUI}
\alias{addEditModuleUI}
\title{Adds the Add/Edit buttons and the associated header header}
\usage{
addEditModuleUI(id)
}
\arguments{
\item{id}{id argument given to module}
}
\description{
Adds the Add/Edit buttons and the associated header header
}
\examples{
ui <- fluidPage(
addEditModuleUI("data")
)

server <- function(input, output, session) {

}

shiny::shinyApp(ui, server)

}
