% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditModules.R
\name{addModalUI}
\alias{addModalUI}
\title{Create Modal: UI function}
\usage{
addModalUI(id, modalTitle)
}
\arguments{
\item{id}{character name for the namespace of the module}

\item{modalTitle}{Character string for title to be displayed at the top of
the modal.}
}
\description{
This function and \code{\link{addModal}} are used in conjunction to create
the UI and server elements necessary to control the modal
}
