% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditModules.R
\name{addModuleUI}
\alias{addModuleUI}
\title{Add database entry button: UI function}
\usage{
addModuleUI(id)
}
\arguments{
\item{id}{character name for the namespace of the module}
}
\value{
A shiny \code{\link[shiny]{actionButton}} which opens a modal to
  allow user to input data to be added to a database
}
\description{
This function and \code{\link{addModule}} are used in conjunction to add the
UI and server elements necessary to add a row to a given table in a database
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage(
    addModuleUI("data")
  )

  server <- function(input, output, session) {

  }

  shinyApp(ui, server)
}

}
\seealso{
\code{\link{addModule}}
}
