% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtModule.R
\name{applyFilters}
\alias{applyFilters}
\title{Apply filters to data displayed in dtModule}
\usage{
applyFilters(.data, x, input)
}
\arguments{
\item{.data}{the dataframe to be filtered and displayed in the dtModule}

\item{x}{a data frame row from filterData data.frame.}

\item{input}{These parameters are handled by
\code{\link[shiny]{callModule}} and can be ignored.}
}
\description{
Apply filters to data displayed in dtModule
}
