% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditModules.R
\name{callModalUI}
\alias{callModalUI}
\title{Call modalUI function}
\usage{
callModalUI(modalUI, ..., session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{modalUI}{Function that creates the modal UI.}

\item{...}{Additional parameters to mass to modalUI function

list(
  selectizeInput(ns("Species"), "Species",
                 choices = reactiveData$flowers$flowerName),
  selectizeInput(ns("smell"), "Smell", choices = irisStaticChoices$smell)
)}

\item{session}{These parameters are handled by
\code{\link[shiny]{callModule}} and can be ignored.}
}
\description{
This function wraps around the the modalUI argument to call the function
withing the modal server functions
}
