% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectChoices.R
\name{choicesReactive}
\alias{choicesReactive}
\title{Create reactive for select/selectize input choices}
\usage{
choicesReactive(inputData, reactiveData, staticChoices = NULL)
}
\description{
Create reactive for select/selectize input choices
}
