% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtModule.R
\name{dtFilterUpdates}
\alias{dtFilterUpdates}
\title{Update filter choices}
\usage{
dtFilterUpdates(input, output, session, filterData, reactiveData)
}
\arguments{
\item{input}{These parameters are handled by
\code{\link[shiny]{callModule}} and can be ignored.}

\item{output}{These parameters are handled by
\code{\link[shiny]{callModule}} and can be ignored.}

\item{session}{These parameters are handled by
\code{\link[shiny]{callModule}} and can be ignored.}

\item{filterData}{Hello}

\item{reactiveData}{Reactive which stores all of the tables from the database
as seperate \code{data.frames}}
}
\description{
Update filter choices
}
