% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtModule.R
\name{dtModule}
\alias{dtModule}
\title{Add reactive datatable: server function}
\usage{
dtModule(input, output, session, reactiveData, dbTable, filterData = NULL)
}
\arguments{
\item{filterData}{hello}
}
\description{
Add reactive datatable: server function
}
