% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtModule.R
\name{dtModuleUI}
\alias{dtModuleUI}
\title{Add reactive datatable: UI function}
\usage{
dtModuleUI(id, filterData = NULL)
}
\arguments{
\item{id}{character name for the namespace of the module}

\item{filterData}{hello}
}
\description{
Add reactive datatable: UI function
}
