% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditModules.R
\name{editModalUI}
\alias{editModalUI}
\title{Create Edit Modal: UI function}
\usage{
editModalUI(id, editTitle)
}
\arguments{
\item{id}{character name for the namespace of the module}

\item{editTitle}{Title to display at the top of the edit Modal}
}
\description{
This function and \code{\link{editModal}} are used in conjunction to create
the UI and server elements necessary to control the modal
}
