% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCallbacks.R
\name{insertCallback}
\alias{insertCallback}
\title{Insert data into database}
\usage{
insertCallback(ids, db, tab, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{ids}{Character vector of shiny input ids to be compiled and added to
database.}

\item{db}{a \code{\link[DBI:DBIConnection-class]{DBIConnection}} object, as
returned by \code{\link[DBI]{dbConnect}}. In other words, the object the
database connection is saved to.}

\item{tab}{The table in the database the inputs will be added to.}

\item{session}{These parameters are handled by
\code{\link[shiny]{callModule}} and can be ignored.}
}
\description{
\code{insertCallback} takes shiny input values supplied by the user, collects
them into a \code{data.frame} and then appends them into a database
}
