% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run example app using all functionality}
\usage{
runExample()
}
\description{
This app uses the \code{iris} and \code{mtcars} data to showcase the
\code{CARTMod} modules and functions in action.
}
