% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectChoices.R
\name{valueLabel}
\alias{valueLabel}
\title{Define label/value pairs for select/selectize inputs}
\usage{
valueLabel(df, value, label)
}
\arguments{
\item{df}{A \code{data.frame} where the choices for the select/selectize
input are. (note: \code{data.frame} used in this argument should live in
the reactiveData object passed from the
\code{\link[CARTMod]{choicesReactive}} function)}

\item{value}{The column in \code{df} where the numeric ID value for a choice
is stored.}

\item{label}{The column in \code{df} where the charater label/name for a
choice is stored.}
}
\description{
Define label/value pairs for select/selectize inputs
}
