% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditModules.R
\name{addModalUI}
\alias{addModalUI}
\title{Create Add Modal: UI function}
\usage{
addModalUI(id, addTitle)
}
\arguments{
\item{id}{character name for the namespace of the module}

\item{addTitle}{Title to display at the top of the add Modal}
}
\description{
This function and \code{\link{addModal}} are used in conjunction to create
the UI and server elements necessary to control the modal
}
