% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditModules.R
\name{addModule}
\alias{addModule}
\title{Add database entry button: server function}
\usage{
addModule(
  input,
  output,
  session,
  modalTitle,
  inputData,
  db,
  dbTable,
  reactiveData,
  checkDuplicate = NULL,
  additionalInputs = NULL
)
}
\arguments{
\item{input, output, session}{These parameters are handled by
\code{\link[shiny]{callModule}} and can be ignored.}

\item{modalTitle}{Character string for title to be displayed at the top of
the modal.}

\item{inputData}{a \code{data.frame} containing columns \code{ids, labels,
  type, choicesTable, choicesValues, choicesLabels} which correspond to
  \describe{
    \item{\code{ids}}{The field names in the database. These ids will also be
    used as the inputId for shiny inputs.}

    \item{\code{labels}}{These are character values which will be used as the
    label for shiny inputs and be displayed above the input in the UI.}

    \item{\code{type}}{The type of shiny input to be used.}

    \item{\code{choicesTable, choicesValues, choicesLabels}}{Only used if the
    desired input has predefined choices
    (\code{\link[shiny:selectInput]{selectInput/selectizeInput})}. If neither
    of these inputs are used, then the values should be \code{NA}. If one of
    these inputs is used then:
      \describe{
        \item{\code{choicesTable}}{Is the table in the database where the
        choices for the given input are.}
        \item{\code{choicesValues}}{Is the column in the table which stores
        the numeric identifier for the choice}
        \item{\code{choicesLabels}}{Is the column in the table which stores
        the character name for the choices}
      }
    }
  }}

\item{db}{a \code{\link[DBI:DBIConnection-class]{DBIConnection}} object, as
returned by \code{\link[DBI]{dbConnect}}. In other words, the object the
database connection is saved to.}

\item{dbTable}{The database table the new data will be added to.}

\item{reactiveData}{Reactive which stores all of the tables from the database
as seperate \code{data.frames}}

\item{checkDuplicate}{Character vector of columns names (corresponds to input
ids) that should be checked in the database for duplication. Default value
is \code{NULL} meaning no variables are checked for duplication}

\item{additionalInputs}{Additional shiny inputs to be dislayed in the
inputModal provided as a list.}
}
\value{
Shiny \code{\link[shiny]{observeEvent}}'s which control actions when
  the add button is pressed, as well as the save button in the modal.
}
\description{
This function and \link{addModuleUI} are used in conjection to add the UI and
server elements necessary to add a row to a given table in a database
}
\seealso{
\code{\link{addModuleUI}}
}
