% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtModule.R
\name{dtModule}
\alias{dtModule}
\title{Create and display datatable: server function}
\usage{
dtModule(
  input,
  output,
  session,
  reactiveData,
  dbTable,
  filterData = NULL,
  staticChoices
)
}
\arguments{
\item{input}{These parameters are handled by
\code{\link[shiny]{callModule}} and can be ignored.}

\item{output}{These parameters are handled by
\code{\link[shiny]{callModule}} and can be ignored.}

\item{session}{These parameters are handled by
\code{\link[shiny]{callModule}} and can be ignored.}

\item{reactiveData}{Reactive which stores all of the tables from the database
as seperate \code{data.frames}}

\item{dbTable}{The database table the new data will be added to.}

\item{filterData}{hello!}

\item{staticChoices}{\code{list} where choices for static selectize
inputs are defined.}
}
\description{
Create and display datatable: server function
}
