% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtModule.R
\name{dtModuleUI}
\alias{dtModuleUI}
\title{Add reactive datatable: UI function}
\usage{
dtModuleUI(id)
}
\arguments{
\item{id}{character name for the namespace of the module}
}
\description{
Add reactive datatable: UI function
}
