% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCallbacks.R
\name{insertCallback}
\alias{insertCallback}
\title{Insert data into database}
\usage{
insertCallback(
  inputData,
  db,
  dbTable,
  session = shiny::getDefaultReactiveDomain()
)
}
\description{
\code{insertCallback} takes shiny input values supplied by the user, collects
them into a \code{data.frame} and then appends them into a database
}
