% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditModules.R
\name{modalInputs}
\alias{modalInputs}
\title{Create list of shiny inputs for modal}
\usage{
modalInputs(
  ns = ns,
  inputData,
  values,
  choices,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{ns}{namespace function passed from calling environment}

\item{inputData}{a \code{data.frame} containing columns \code{ids, labels,
  type, choicesTable, choicesValues, choicesLabels} which correspond to
  \describe{
    \item{\code{ids}}{The field names in the database. These ids will also be
    used as the inputId for shiny inputs.}

    \item{\code{labels}}{These are character values which will be used as the
    label for shiny inputs and be displayed above the input in the UI.}

    \item{\code{type}}{The type of shiny input to be used.}

    \item{\code{choicesTable, choicesValues, choicesLabels}}{Only used if the
    desired input has predefined choices
    (\code{\link[shiny:selectInput]{selectInput/selectizeInput})}. If neither
    of these inputs are used, then the values should be \code{NA}. If one of
    these inputs is used then:
      \describe{
        \item{\code{choicesTable}}{Is the table in the database where the
        choices for the given input are.}
        \item{\code{choicesValues}}{Is the column in the table which stores
        the numeric identifier for the choice}
        \item{\code{choicesLabels}}{Is the column in the table which stores
        the character name for the choices}
      }
    }
  }}

\item{values}{Optional argument to be used when the inputs are being
populated from an observation in the database. (NEED MORE DOCUMENTATION
HERE ONCE THE EDIT FUNCTIONALITY IS BUILT OUT)}

\item{choices}{Optional argument to provide the choices for
\code{\link[shiny]{selectInput}} and
\code{\link[shiny:selectInput]{selectizeInput}} inputs.}

\item{session}{These parameters are handled by
\code{\link[shiny]{callModule}} and can be ignored.}
}
\description{
Takes a \code{data.frame} containing information about shiny inputs and
additional optional parameters to create list of shiny inputs to be displayed
in a modal
}
