% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectChoices.R
\name{valueLabel}
\alias{valueLabel}
\title{Define label/value pairs for select/selectize inputs}
\usage{
valueLabel(df, value, label)
}
\arguments{
\item{df}{a data.frame}

\item{value}{value labels}

\item{label}{labels}
}
\description{
Define label/value pairs for select/selectize inputs
}
